requirements.txt
aiofile==3.8.5
aiofiles==23.1.0
alabaster==0.7.13
anyio==3.6.2
argcomplete==2.1.2
arrow==1.2.3
asttokens==2.2.1
asynctest==0.13.0
audioread==3.0.0
autopep8==2.0.2
b2==3.8.0
b2sdk==1.20.0
Babel==2.12.1
backblaze==0.1.1
backcall==0.2.0
beautifulsoup4==4.12.2
caio==0.9.12
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
css-html-js-minify==2.5.5
cycler==0.11.0
decorator==5.1.1
distlib==0.3.6
dnspython==2.3.0
docutils==0.19
executing==1.2.0
filelock==3.11.0
Flask==2.2.3
fonttools==4.39.0
gunicorn==20.0.4
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
imagesize==1.4.1
ipython==8.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.1
librosa==0.10.0
llvmlite==0.39.1
logfury==1.0.1
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
msgpack==1.0.5
numba==0.56.4
numpy==1.23.5
packaging==23.0
pandas==1.5.3
parso==0.8.3
phx-class-registry==4.0.5
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
pooch==1.7.0
praat-parselmouth==0.4.3
prompt-toolkit==3.0.38
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pydub==0.25.1
Pygments==2.14.0
pymongo==3.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2022.7.1
requests==2.28.2
rst2ansi==0.1.5
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
soundfile==0.12.1
soupsieve==2.4
soxr==0.3.4
Sphinx==6.1.3
sphinx-material==0.0.35
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib-trio==1.1.2
stack-data==0.6.2
tabulate==0.9.0
text-unidecode==1.3
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
Unidecode==1.3.6
urllib3==1.26.15
virtualenv==20.21.0
wcwidth==0.2.6
Werkzeug==2.2.3
